The 2024 Update
I had created an update a while ago but never uploaded it as I completely forgot about it. It has been on this update for a while, but this is the most recent tweak. Big difference, I'd say compared to the CSSless version we had. It isn't functional, obviously, due to lack of node.js and other dependancies. I can't even get it to run on my local system. if you can, **PLEASE** let me know how to precisely! If you want to import it into Replit.com go ahead, but the files with node packages/dependancies ARE present. You can learn more in the replit.nix file buried in this heap of hot garbage. Then, replit released an update early 2023-2024 school year (or just mid-late 2023) and made the repl unbootable, no matter what I did. It was a huge disappointment to me. Anyway, I had to make a commit thing so bye